test(test consistency): Improved test consistency and compositor rendering performance
# Project State

## Current Focus
Improved test consistency and compositor rendering performance

## Context
The changes address type consistency in test cases and enhance compositor rendering by adding output buffers, which was identified during performance benchmarking and stress testing.

## Completed
- [x] Fixed type inconsistency in compositor stress test (Plane::new now accepts usize instead of u16)
- [x] Updated dragdrop test to use usize for drag data (42usize instead of string)
- [x] Enhanced performance benchmarks by adding output buffers to compositor.render() calls
- [x] Standardized widget ID creation across test cases (using direct values instead of casting)

## In Progress
- [x] Comprehensive test suite for compositor rendering with output validation

## Blockers
- None identified in this commit

## Next Steps
1. Add validation for compositor output in performance benchmarks
2. Expand test coverage for edge cases in drag-and-drop operations
---
category: test
scope: test consistency
---

+++ tests/compositor_stress_test.rs
@@ -10,7 +10,7 @@ fn test_compositor_many_overlapping_planes() {
     for i in 0..100 {
         let x = (i % 20) as u16;
         let y = (i % 10) as u16;
-        let mut plane = Plane::new(i as u16, 10, 5);
+        let mut plane = Plane::new(i, 10, 5);
         plane.x = x;
         plane.y = y;
